[Fix] 오류해결

+++ back/src/routers/commentRouter.js
@@ -54,7 +54,7 @@ commentRouter.get(
       const page = +req.query.page;
       const perPage = +req.query.perPage;
 
-      const comments = await commentService.getComments({
+      const comments = await commentService.getCommentByPostId({
         postId, 
         page, 
         perPage,
@@ -86,14 +86,20 @@ commentRouter.put(
   commentValidate.updateComment,
   async (req, res, next) => {
     try {
-      //const userId = req.currentUserId;
+      const userId = req.currentUserId;
       const commentId = req.params.id;
+      const comment = await commentService.getCommentById(commentId);
+
+      if(userId !== comment.userId) {
+        const error = new Error("수정 권한이 없습니다.")
+        throw error;
+      }
+
       const content = req.body.content ?? null;
       
       const toUpdate = { content };
-
       const updateComment = await commentService.setComment({ commentId, toUpdate });
-
+      
       const body = {
         success: true,
         comment: updateComment
@@ -114,7 +120,16 @@ commentRouter.delete(
   login_required,
   async (req, res, next) => {
     try {
+      const userId = req.currentUserId;
       const commentId = req.params.id;
+      const comment = await commentService.getCommentById(commentId);
+
+      if(userId !== comment.userId) {
+        const error = new Error("삭제 권한이 없습니다.")
+        throw error;
+      }
+
+
       const isDeleted = await commentService.deleteComment(commentId);
 
       res.status(200).json(isDeleted);
