+++ back/src/routers/diaryRouter.js
@@ -85,7 +85,15 @@ diaryRouter.put(
   diaryValidate.updateDiary,
   async (req, res, next) => {
   try {
+    const userId = req.currentUserId;
     const diaryId = req.params.id;
+    
+    const diary = await diaryService.getDiaryById(diaryId);
+      
+    if(userId !== diary.userId) {
+      const error = new Error("수정 권한이 없습니다.")
+      throw error;
+    }
 
     const imageURL = req.body.imageURL ?? null;
     const content = req.body.content ?? null;
@@ -115,7 +123,16 @@ diaryRouter.put(
  */
 diaryRouter.delete("/diaries/:id", login_required, async (req, res, next) => {
   try {
+    const userId = req.currentUserId;
     const diaryId = req.params.id;
+
+    const diary = await diaryService.getDiaryById(diaryId);
+      
+    if(userId !== diary.userId) {
+      const error = new Error("삭제 권한이 없습니다.")
+      throw error;
+    }
+
     const isDeleted = await diaryService.deleteDiary(diaryId);
 
     res.status(200).json(isDeleted);
